backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
gunicorn>=21.2.0
python-dotenv==1.0.0
google-genai
opencv-python-headless>=4.9.0
Pillow>=10.0.0
reportlab>=4.0.0
numpy>=1.26.0
SpeechRecognition>=3.10.0
pydub>=0.25.0
